refactor(lib/tui): move "draw_area_*" logic to TUI utils
as the tui is the only place it is used
